refactor(page): extract pure helpers/types into modules (phase 1, batches 1–2)
Begin splitting the 23k-line src/app/page.tsx by moving pure
module-level helpers, types, and constants into focused feature/util
modules. Functions moved verbatim; call sites unchanged (re-imported
by the same name), so no behavior change.

Extracted:
- utils/base64, utils/geometrySnapshot, utils/geometryScaling, utils/jsonFields
- features/hole-punching/{holePunchGeometry,holePunchPersistence}
- features/hollowing/{hollowingSerialize,hollowingPreviewTypes,hollowingPreviewCache}
- features/printing/printingMonitorFormat

page.tsx: 23,133 -> 22,483 lines. Verified via `tsc --noEmit` (zero
errors in any touched file).

Also support the in-progress pnpm migration so types resolve under
pnpm's strict layout: declare @types/three, zustand, three-mesh-bvh;
pin three/manifold-3d/@react-three/fiber/@react-three/drei (+zustand,
three-mesh-bvh) to the package-lock.json versions; switch
src/utils/bvh.ts to the bare `three-mesh-bvh` specifier. (package.json
also carries pre-existing npm->pnpm script edits from this branch.)
